Adds svgr to boost, removes Storybook default SVG
Storybook comes with a default SVG (and all other assets loader).
The simple hack right now is to just modify the default loader
regex and stick our loader in.

In the future it would probably nice to put this functionality
into a Storybook preset:
https://github.com/storybookjs/storybook/blob/next/docs/src/pages/presets/writing-presets/index.md
and give back to the OS community. Could be a good first
contribution, since we're relying heavily on Storybook.