feat: Read Boost Theme theme prperties from project.yml
With this commit we generate settings files for SCSS and JS from the
layout configuration inside the project configuration. We optionally use
those created settings files for the default theme in the boost theme.

This way we can change the theme from the project configuration as a
central place.

Bonus point: The breakpoint configuration is also available in JS &
SCSS.